requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
celery==5.5.0
redis==4.5.1
requests==2.31.0
python-dotenv==1.0.0